+++ core/arch/arm/plat-stm32mp1/shared_resources.c
@@ -15,7 +15,6 @@
 #include <kernel/boot.h>
 #include <kernel/panic.h>
 #include <kernel/pm.h>
-#include <kernel/spinlock.h>
 #include <libfdt.h>
 #include <mm/core_memprot.h>
 #include <platform_config.h>
@@ -28,40 +27,6 @@
  * new resources. This ensures resource state cannot change.
  */
 static bool registering_locked;
-
-/*
- * Shared register access: upon shared resource lock
- */
-static unsigned int shregs_lock = SPINLOCK_UNLOCK;
-
-/* Shared resource lock: assume not required if MMU is disabled */
-static uint32_t lock_stm32shregs(void)
-{
-	return may_spin_lock(&shregs_lock);
-}
-
-static void unlock_stm32shregs(uint32_t exceptions)
-{
-	may_spin_unlock(&shregs_lock, exceptions);
-}
-
-void io_mask32_stm32shregs(vaddr_t va, uint32_t value, uint32_t mask)
-{
-	uint32_t exceptions = lock_stm32shregs();
-
-	io_mask32(va, value, mask);
-
-	unlock_stm32shregs(exceptions);
-}
-
-void io_clrsetbits32_stm32shregs(vaddr_t va, uint32_t clr, uint32_t set)
-{
-	uint32_t exceptions = lock_stm32shregs();
-
-	io_clrsetbits32(va, clr, set);
-
-	unlock_stm32shregs(exceptions);
-}
 
 /*
  * Shared peripherals and resources registration
